requirements.txt
PyYAML
mistralai
streamlit
langchain
langchain_openai
langchain_mistralai
manim
